+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog since 2.3.0
 
 * Fixed blocksCreate, blocksDestroy, usersLookup, usersShow, usersSearch to reflect version 1.1 of the API.
+* Implemented usersContributees, usersContributors.
 
 # Changelog since 2.2.1